fix: prevent radio lifecycle crashes

+++ src/include/radio_utils.hpp
@@ -9,7 +9,7 @@ uint64_t RadioCurrentTimeMillis();
 	if (source == 0) {                                                                                                 \
 		FlatVector::SetNull(destination, 0, true);                                                                     \
 	} else {                                                                                                           \
-		FlatVector::GetData<uint64_t>(destination)[0] = source;                                                        \
+		FlatVector::GetData<int64_t>(destination)[0] = static_cast<int64_t>(source);                                   \
 	}
 
 } // namespace duckdb
